pyproject.toml
[tool.poetry]
name = "strawberry-django-test"
version = "0.1.0"
description = ""
authors = ["Jonathan Kim <jkimbo@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1"
strawberry-graphql = { path = "../strawberry/", develop = true }

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.3"
flake8-bugbear = "^20.1.4"
flake8-black = "^0.2.1"
pdbpp = "^0.10.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
